requirements.txt
pandas
numpy
tabulate
python-dotenv
unsloth
unsloth_zoo
transformers
datasets
trl
bitsandbytes
xformers
openai
llama-cpp-python
gguf
